[서버] 오류 수정 및 기술 문서를 위한 파일 추가

+++ server/src/domain/Review.ts
@@ -19,7 +19,7 @@ export default class Review extends BaseEntity {
     @PrimaryGeneratedColumn('increment')
     public id?: number;
 
-    @Column('int')
+    @Column('tinyint')
     public score?: number;
 
     @Column('text')
